+++ tests/testthat/test-ballot_io_deprecated.R
@@ -1,14 +1,22 @@
 test_that("read_ballots and write_ballots are inverse.", {
-  bs <- ranked_ballots(list(LETTERS[1:5], LETTERS[2:4], LETTERS[2:4]))
-  expect_true(identical(
+  bs <- suppressWarnings(
+    ranked_ballots(list(LETTERS[1:5], LETTERS[2:4], LETTERS[2:4]))
+  )
+  expect_warning({
+    out <- write_ballots(bs, return_lines = TRUE, suppress = TRUE)
+  })
+  expect_warning({
+    read_ballots(out)
+  })
+  expect_true(suppressWarnings(identical(
     bs,
     read_ballots(write_ballots(bs, return_lines = TRUE, suppress = TRUE))
-  ))
+  )))
 
   # Ensure all candidates show in the output
   for (l in LETTERS[1:5]) {
     expect_output(
-      write_ballots(bs),
+      suppressWarnings(write_ballots(bs)),
       l
     )
   }
@@ -21,10 +29,10 @@ test_that("read_ballots and write_ballots are inverse.", {
     "(B, C, D) : 2"
   )
   expect_true(all(
-    lines == write_ballots(
+    lines == suppressWarnings(write_ballots(
       read_ballots(lines),
       return_lines = TRUE,
       suppress = TRUE
-    )
+    ))
   ))
 })