+++ tests/testthat/test-minmaxdepth.R
@@ -9,16 +9,24 @@ test_that(paste0(
     max_depth = 3,
     vd = FALSE
   )
-  bs <- ranked_ballots(c("C", "B", "A", "D", "E"))
+  bs <- prefio::preferences(
+    matrix(
+      c(3, 2, 1, 4, 5),
+      ncol = 5,
+      byrow = TRUE
+    ),
+    format = "ranking",
+    item_names = LETTERS[1:5]
+  )
 
   update(dtree, bs)
 
   test_sample <- sample_predictive(dtree, 1)
 
-  expect_true(all(test_sample[[1]][1:3] == bs[[1]][1:3]))
+  expect_true(test_sample$preferences == bs[, 1:3, by.rank = TRUE])
 })
 
-test_that("sample_predictive produces valid ballots", {
+test_that("sample_predictive produces valid-length ballots", {
   dtree <- dirtree(
     candidates = LETTERS[1:10],
     a0 = 1.,
@@ -27,10 +35,10 @@ test_that("sample_predictive produces valid ballots", {
     vd = FALSE
   )
 
-  test_samples <- sample_predictive(dtree, 1000)
+  test_samples <- sample_predictive(dtree, 1000)$preferences
 
-  expect_true(all(lapply(test_samples, length) >= 2))
-  expect_true(all(lapply(test_samples, length) <= 8))
+  expect_true(all(rowSums(!is.na(as.matrix(test_samples))) >= 2))
+  expect_true(all(rowSums(!is.na(as.matrix(test_samples))) <= 8))
 })
 
 test_that("sample_predictive produces correct number of ballots when a0=0", {
@@ -47,7 +55,7 @@ test_that("sample_predictive produces correct number of ballots when a0=0", {
   dtree$a0 <- 0
   dtree$max_depth <- 7
 
-  expect_equal(length(sample_predictive(dtree, 1000)), 1000)
+  expect_equal(sum(sample_predictive(dtree, 1000)$frequencies), 1000)
 })
 
 test_that("min_depth cannot be less than max_depth.", {